use default_real_time_data for initial settings
In a new install where some default values are
needed for real_time_data, a new file default_real_time_data
is used.
If a file new_install_indicator doesn't exist in data path,
it is considered as a new_install.
Then, default_real_time_data is copied into real_time_data.

The problem with only using real_time_data is that it's
hard to deal with merge conflicts when real_time_data is
updated in base repo.